Fix cast int

+++ srcs/server/ft_irc_cmd_leave.c
@@ -6,7 +6,7 @@
 /*   By: gtorresa <gtorresa@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/08 00:16:36 by gtorresa          #+#    #+#             */
-/*   Updated: 2017/11/13 12:37:23 by gtorresa         ###   ########.fr       */
+/*   Updated: 2017/11/13 12:38:38 by gtorresa         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -88,9 +88,9 @@ void			ft_irc_leave(t_env *e, int cs, char *chan)
 		l = l->next;
 	}
 	e->chanel = ft_irc_clear_empty_chanel(e->chanel);
-	loop = ft_lstlen(e->fds[cs].chan_user);
+	loop = (int)ft_lstlen(e->fds[cs].chan_user);
 	e->fds[cs].chan_user = ft_irc_chan_user_del(e->fds[cs].chan_user, chan);
-	if (loop != ft_lstlen(e->fds[cs].chan_user))
+	if (loop != (int)ft_lstlen(e->fds[cs].chan_user))
 		ft_irc_cmd_leave_print(e, cs, chan);
 	if (e->fds[cs].chanel)
 		free(e->fds[cs].chanel);
